1. 修改脚本

+++ tools.sh
@@ -15,8 +15,12 @@ start() {
 }
 
 stop() {
-    ps -ef | grep python | grep bot.py | awk '{print $2}' | xargs kill -9
-    echo "stop zbot complete."
+    if ps -ef | grep python | grep bot.py > /dev/null
+        ps -ef | grep python | grep bot.py | awk '{print $2}' | xargs kill -9
+        echo "stop zbot complete."
+    else
+        echo "zbot no found."
+    fi
 }
 
 update() {
